Update README.md
Honestly, I'm confused by this proposal. Is your question about revenue or the customer journey? What data are you going to use to answer these questions? The six-step process is a generic DS process and uninformative---I really don't know what you will be working on after reading this. Step six suggests that you are creating a dashboard for a CMO. Is that the product? You might want to consider using the template I supplied to revise and prepare for the next round. I need to see more specifics and less jargon---e.g., what is an "executive level insight?"

80/100